app/requirements.txt
fastapi
uvicorn 
motor
python-dotenv